json_traverse.rs test cleanup
