+++ crates/document/src/json_traverse.rs
@@ -102,198 +102,3 @@ where
 
     Ok(())
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use serde_json::json;
-
-    #[test]
-    fn test_traverse_simple_object() {
-        let json = json!({"height": 168, "weight": 70});
-        let mut results = Vec::new();
-
-        traverse_json(
-            &json,
-            |path, value| {
-                results.push((path.clone(), value.clone()));
-                Ok(())
-            },
-            &index_traverse_options(),
-        )
-        .unwrap();
-
-        assert_eq!(results.len(), 2);
-        // Results may be in any order due to HashMap iteration
-    }
-
-    #[test]
-    fn test_traverse_nested_object() {
-        let json = json!({"custom": {"height": 168}});
-        let mut results = Vec::new();
-
-        traverse_json(
-            &json,
-            |path, value| {
-                results.push((path.clone(), value.clone()));
-                Ok(())
-            },
-            &index_traverse_options(),
-        )
-        .unwrap();
-
-        assert_eq!(results.len(), 1);
-        let (path, value) = &results[0];
-        assert_eq!(path.len(), 2);
-        assert_eq!(*value, json!(168));
-    }
-
-    #[test]
-    fn test_traverse_array() {
-        let json = json!({"tags": ["a", "b", "c"]});
-        let mut results = Vec::new();
-
-        traverse_json(
-            &json,
-            |path, value| {
-                results.push((path.clone(), value.clone()));
-                Ok(())
-            },
-            &index_traverse_options(),
-        )
-        .unwrap();
-
-        assert_eq!(results.len(), 3);
-        // Each element should have path [Property("tags"), Index]
-        for (path, _) in &results {
-            assert_eq!(path.len(), 2);
-        }
-    }
-
-    #[test]
-    fn test_traverse_empty_object() {
-        let json = json!({});
-        let mut results = Vec::new();
-
-        traverse_json(
-            &json,
-            |path, value| {
-                results.push((path.clone(), value.clone()));
-                Ok(())
-            },
-            &index_traverse_options(),
-        )
-        .unwrap();
-
-        // Empty object produces no entries
-        assert_eq!(results.len(), 0);
-    }
-
-    #[test]
-    fn test_traverse_empty_array() {
-        let json = json!([]);
-        let mut results = Vec::new();
-
-        traverse_json(
-            &json,
-            |path, value| {
-                results.push((path.clone(), value.clone()));
-                Ok(())
-            },
-            &index_traverse_options(),
-        )
-        .unwrap();
-
-        // Empty array produces no entries
-        assert_eq!(results.len(), 0);
-    }
-
-    #[test]
-    fn test_traverse_null() {
-        let json = json!(null);
-        let mut results = Vec::new();
-
-        traverse_json(
-            &json,
-            |path, value| {
-                results.push((path.clone(), value.clone()));
-                Ok(())
-            },
-            &index_traverse_options(),
-        )
-        .unwrap();
-
-        assert_eq!(results.len(), 1);
-        assert!(results[0].0.is_empty());
-        assert!(results[0].1.is_null());
-    }
-
-    #[test]
-    fn test_traverse_scalar_types() {
-        let test_cases = vec![
-            json!(true),
-            json!(false),
-            json!(42),
-            json!(3.15),
-            json!("hello"),
-            json!(null),
-        ];
-
-        for json in test_cases {
-            let mut results = Vec::new();
-            traverse_json(
-                &json,
-                |path, value| {
-                    results.push((path.clone(), value.clone()));
-                    Ok(())
-                },
-                &index_traverse_options(),
-            )
-            .unwrap();
-
-            assert_eq!(results.len(), 1);
-            assert!(results[0].0.is_empty());
-        }
-    }
-
-    #[test]
-    fn test_traverse_nested_array() {
-        let json = json!({"data": [[1, 2], [3, 4]]});
-        let mut results = Vec::new();
-
-        traverse_json(
-            &json,
-            |path, value| {
-                results.push((path.clone(), value.clone()));
-                Ok(())
-            },
-            &index_traverse_options(),
-        )
-        .unwrap();
-
-        // Should visit leaf scalars: 1, 2, 3, 4
-        assert_eq!(results.len(), 4);
-    }
-
-    #[test]
-    fn test_traverse_mixed_array() {
-        let json = json!([1, "a", true, null]);
-        let mut results = Vec::new();
-
-        traverse_json(
-            &json,
-            |path, value| {
-                results.push((path.clone(), value.clone()));
-                Ok(())
-            },
-            &index_traverse_options(),
-        )
-        .unwrap();
-
-        assert_eq!(results.len(), 4);
-        // Each has Index in path
-        for (path, _) in &results {
-            assert_eq!(path.len(), 1);
-        }
-    }
-}
